Fix preferred_python_version for RedHat 7
Set preferred_python_version by having the per-platform code add
an extra key to the "image" dict, and then having common code read it and
deal with defaults as necessary. This replaces two different code paths,
both of which were vulnerable to having their changes overwritten by
the later code intended to ensure a default was set.

Remove some code to deal with ancient versions (BDR1), which could
change image information later in the process and so break detection of
preferred_python_version.