Remove $umask directive from rsyslog configuration
We have always previously set a umask of 0000 in ryslog configuration
(specifically, in the configuration we supply for postgres logs, but the
parameter acts universally). But this results in /run/rsyslogd.pid being
created world-readable, which is a problem for any customer with a
general prohibition on world-readable files. We can simply remove the
directive and let the customer's preferred system-wide setting apply.