Revert "Fix the condition for setting camo configs tpa-212"
This reverts 09d2655 because 'bdr_version' does not carry the same
information as 'bdr_version_num' so the change in tests does not give
us the expected result. We will need to find a different way to figure
out BDR version for clusters that are going through fresh deployments
and don't yet have a running database service.