Remove empty environment definition in raw modules
Remove the empty environment definition in raw module that were
added to silence the warnings in the first place.
this stopped working when switching to community ansible,
this Fixes the warning in both situation where target_environment
is defined and when it's not.
Only one occurence can't be fixed in architectures/lib/init.yml