Skip some tasks if we're running in a container
Per Support ticket 84721, customer needs to deploy to Docker containers
using the bare platform, because this allows them to have a development
environment that is closer to the environment they will deploy to their
customers, with minimal changes to the `config.yml` file.

We invoke systemd-detect-virt from library/minimal_setup to detect if we
are running inside a container, and set target_container_type if that's
the case. We add conditions based on this new fact to avoid some tasks:

- Tasks with the "hostname" tag: The hostname is changed with systemd,
  and we are already doing this only for "platform != docker".

- Tasks with the "sysctl" tag: We can't change sysctl settings inside a
  container, and we are already doing this only for "platform !=
  docker".

- Tasks under the "sysctl" directory that are not directly tagged with
  the "sysctl" tag.

- Tasks under "os/Debian" and "os/RedHat", because we can't configure
  grub inside containers.

With that, the customer will not need to "--skip-tags hostname,sysctl"
anymore, because essentially the tasks under the tags "hostname" and
"sysctl" will be automatically skipped, regardless of if the customer is
using "platform: docker" or not.

References: TPA-156

Signed-off-by: William Ivanski <william.ivanski@gmail.com>

With contributions from John.