Accept ip addresses in hostnames file
We extend the format of the hostnames file to allow an optional ip
address after each hostname. If an address is given, it will be added
to the ip_address field for that host in config.yml .

Additionally, hostnames are no longer randomised by default if a
non-default hostnames file is given.