Add bdr-pre-group-join hook
Per RT86745, AutoPartition is not working because it checks for
partition metadata on the `RAFT_LEADER`. If the `RAFT_LEADER` is the
empty witness as configured by TPAexec, then AutoPartition fails.

A workaround is to have the `witness-only` replication set configured to
automatically add tables, as well as a DDL Filter replicating all DDL
associated with the replication set.

However, this can't be configured as a `post-deploy.yml` hook, because
at this point the witness will be already joined and empty. The previous
hook, `bdr-post-group-creation.yml`, is triggered before the
`witness-only` replication set is created, so we also can't use this
hook.

So with this commit we provide add a new hook called
`bdr-pre-group-join`, which is executed at the end of `bdrX/init.yml`,
after the replication sets are configured.

In this hook, in file `hooks/bdr-pre-group-join.yml`, we can configure
the witness-only replication set to automatically add tables and add the
DDL filter, which meets the workaround described above:

```
- name: Configure the witness-only repset to automatically add tables
  postgresql_query:
    conninfo: "dbname={{ bdr_database }}"
    query: >
      SELECT bdr.alter_replication_set(
        set_name := 'witness-only',
        autoadd_tables := true
      );
  become: true
  become_user: "{{ postgres_user }}"
  when: >
    inventory_hostname == first_bdr_primary

- name: Add a DDL filter to the witness-only repset to replicate all DDL
  postgresql_query:
    conninfo: "dbname={{ bdr_database }}"
    query: >
      SELECT (CASE WHEN
                     NOT EXISTS(
                           SELECT set_ddl_name
                           FROM bdr.ddl_replication
                           WHERE set_ddl_name = 'witness-only'
                     )
                   THEN bdr.replication_set_add_ddl_filter(
                          set_name        := 'witness-only',
                          ddl_filter_name := 'witness-only',
                          command_tag     := NULL::text
                        )
                   ELSE false
              END);
  become: true
  become_user: "{{ postgres_user }}"
  when: >
    inventory_hostname == first_bdr_primary
```

Signed-off-by: William Ivanski <william.ivanski@gmail.com>