Restore code to make all nodes publish to the witness-only replication set
This piece of code was part of the `bdrX/init.yml` file, and upon the
code refactoring that was performed per TPA-193, it was unintentionally
removed.

This task is needed to make sure all nodes publish to the `witness-only`
replication set, and hence the witness will receive DDL and/or BDR admin
functions from other nodes, depending on how the DDL filters are
configured.

Signed-off-by: William Ivanski <william.ivanski@gmail.com>