Remove plain references to hostvars[hostname]
hostvars[hostname] can't be assigned to a variable and later used for
lookups as if it were an ordinary dict, because it's a magic object that
just behaves like a dict when you look up values in it, and the
assignment (sometimes) takes away the magic. This problem doesn't happen
on docker containers or on aws hosts which are using cached inventory,
so it can be hard to reproduce.