Avoid using `hostvars[x]` intermediate variables
Assigning hostvars[x] to a variables makes subsequent expressions easier
to read, but because of hostvars' magical properties, it can also cause
hard-to-diagnose errors, typically manifested as hostvars[x] being seen
as a string instead of a proper dict in which you can lookup keys.

These cases were found by grepping for 'hostvars\[[^\]]*\][^\[\].]' as
part of fixing a similar problem elsewhere.

References: TPA-247