Use consistent git terminology; clean up error handling
Consistently avoid "remote" as a noun to mean "remote repository";
instead, use "repository" or, if there is a risk of confusion, "remote
repository".

At configure-time, if we don't find git, then that isn't an error; but
if we do find it then any failing git commands will raise an exception.

At provision-time, if there are uncommitted changes, fail validation.

When doing git operations at the end of provisioning, ensure we have a
git repository and a configured remote if we are a Tower cluster; add
the files to git if we have a repository, whether we are a Tower cluster
or not; push them only if we are a Tower cluster.

Change git invocations to work with older versions of git back to all
our supported platforms.