Handle per-group 'enable_raft' setting specially for BDRv5
If you set `enable_subgroup_raft: [true|false]` globally, we alter
bdr_node_groups to set `enable_raft` on each entry (if not already
specified). We also exclude this option from alter_node_group_config(),
and handle it using alter_node_group_option() after the group joins have
been completed.

In practice, it should always either be enabled or disabled uniformly
for all subgroups, but tpaexec does not enforce this.

Note that it's not supported to turn off enable_raft once it's set, and
it's not needed to set it on the top group (will always be enabled).

References: TPA-269