TPA-311 Set MSG attribute on exception classes
Address potential conflict with 3rd party Exception classes that define
a message attribute.

In some cases an exception class will have a "message" attribute, which
is more common than the MSG attribute in the BaseTPAException class. We
don't want to import the exception class in configure to do a check
using `isinstance` as we're mitigating failed imports of tpaexec. So the
next best thing is checking for attributes. We could do a dummy check
of exc_value.__bases__ but would that would be recursive and potentially
ugly. Thus, changing the attribute from the more common "message" to
`MSG` is simpler.