Configure YUM repositories for patroni
Do not use EDB managed packages for patroni on RHEL

As haproxy is available in EDB `dl/default` and we need to avoid using
3rd party packages, add an option to the yum config to exclude it. This
is done only for patroni, but there might be an argument to exclude it
in a wider scope, e.g. for any non BDR cluster - as that is the only
legitimate use of the EDB haproxy (although I am not 100% sure on that).

As the version of haproxy in EDB repos has a patch to fix the use of
SCRAM passwords in `pgsql-check` for backend health checks and because
Patroni uses the API endpoints for health checks, explicitly set
`haproxy_backend_queue_check: none`.

Enable extras repo on RHEL 7 for patroni

The extras repo supplies etcd. Some logic is required to determine the
correct file to edit. This is somewhat flimsy so 2 options are provided
to override both the file name and the section (repo name).

   rhel_7_extras_repo_path
   rhel_7_extras_repo_section

An alternative exists to use the pgdg-common-extras repo, however the
etcd package is currently broken on RHEL 7. Config files are overwritten
when the RPM is (re)installed. If that is fixed and the package from
there is preferable this can be reverted.

References: TPA-154, PAT-1