Warn people against using the default t3.micro instances with RH
We've had numerous reports of EC2 t3.micro instances running out of
memory during deployment (typically while running yum). It doesn't
always happen, but it happens enough to be annoying.

Unfortunately, the significant cost difference between t3.micro and
t3.medium, and TPAexec users' propensity to leave abandoned instances
lying around for weeks or months, and the fact that new architectures
often involve a large number of instances means that we can't just use
t3.medium by default (not to mention that micro instances still work
fine on Debian and Ubuntu).

In an ideal world, setting up some sort of lifecycle management so that
we terminate unused instances after a week or so would be the right way
to avoid ballooning costs. We aren't in a position to do that right now,
but we can at least warn the user at configure-time that RH+t3.micro is
a combination that might waste their time.

References: TPA-266