Reformat complex Jinja2 expressions in YAML
We used to do:

    - set_fact:
        thing: "{{
            some complex
            expression
        }}"

This is valid according to the YAML specification (e.g., yamllint
doesn't complain), but the tree-sitter based parser for YAML gets
confused when it sees this construct. Unfortunately, it does not
seem to be easy to fix this.

So now we prefer to use any of these forms:

    - set_fact:
        thing: "{{ otherthing }}"
        thing:
          "{{ longer otherthing }}"
        thing:
          "{{ even longer otherthing
              over multiple
              lines }}"
        thing: >-
          {{
            even longer otherthing
            over multiple lines
          }}
        thing: "{{
            even longer otherthing
            over multiple
            lines
          }}"

Cosmetic for the reader, but helpful for advanced editor functionality.