Disable fallback locations for pgd-proxy for now
This is an advanced feature that we may not want to enable by default,
and may need more than an on/off switch to handle correctly anyway. We
now have the basic mechanism to set/update the fallback group(s), but
it's turned off by default on advice from Suraj.