Set enable_subgroup_raft under cluster_vars
The default was already false, and that's what we set under
cluster_vars, so nothing really changes. This just makes it easier to
see where to turn it on, if you want to try it.

Later, we can figure out how to decide whether to enable it or not
depending on the cluster configuration.