Always set `bdr_node_group: cluster_name` in cluster_vars
Harp and pgd-{cli,proxy} expect the top-level BDR node group name to
match their configured cluster name, so it's easiest to ensure that
the same cluster_name is used everywhere.

One added complication is that cluster_name currently permits characters
that are not valid in BDR group names (especially uppercase characters),
so we have to transform it to a BDR-safe version both in the database
and in the harp/pgd configuration files.

Another complication is that we don't want to change existing clusters.
For one thing, node groups cannot be renamed (their unique ids in the
catalogs are based on a hash of the name). Harp stores cluster metadata
in the DCS indexed by the cluster_name, so you can't just change it and
restart. That means we have to check if cluster.name is already set in
the config, and leave it alone if that's the case. (Note: cluster_name
was never configurable for harp/pgd to begin with, which simplifies
detection.)

References: TPA-297