Introduce harp_manager_user
This commit allows HARP Manager to optionally connect to the local
database using a new user with `bdr_superuser` privileges.

If the user defines `harp_manager_user`, then TPAexec will:

- Create the new user belonging to the `bdr_superuser` role;

- Add the new user password to the Postgres `.pgpass` file of each BDR
node;

- Add a new `local all <user> <method>` entry to the `pg_hba.conf` file
of each BDR node, to allow local connections from the new user.

- Set `bdr_node_local_dsn` (used by HARP Manager) to use the new user.

If `harp_manager_user` is not defined, then TPAexec defaults to the old
behavior of using the `postgres_user`.

We don't change HARP manager user on existing clusters.

On existing clusters `harp_manager_user` is not going to be defined so
in that case, we won't change the user on HARP and postgres files.
However, if `harp_manager_user` is set on config.yml and because this
variable takes precedence, it will be used for new and old clusters.

Also adding `default_harp_manager_user` to the code which will be used
for new clusters if `harp_manager_user` is not set before the deploy
phase.

With valuable contributions from Abhijit and Mario.

References: TPA-309

Co-authored-by: Mario Gonzalez <mario.gonzalez@enterprisedb.com>
Signed-off-by: William Ivanski <william.ivanski@gmail.com>