Avoid repeatedly setting the route_dsn for every node on every deploy
We were treating bdr.node_summary as a summary of bdr.node_config (it
isn't) and checking if the route_dsn column (that isn't) there had a
different value from the route_dsn we wanted to set, and mistakenly
setting it on every deploy because of the mismatch.

Now we switch to bdr.node_config (joined with bdr.node to get the
node_name) and look up node_${option.key} (i.e., not route_dsn but
node_route_dsn instead).