Add a read-only load balancer to haproxy for patroni
A new front end "fe_ro" and a backend "be_ro" provides this "view". The
backend health for patroni uses the /replica endpoint so that traffic
is load balanced over nodes that are otherwise not busy handling writes.
This matches the recommended configuration layout in documentation
provided to customers from the patroni team.

References: TPA-154, PAT-1