Increase attempts to check for pgreceive_wal/xlog
In recent CI runs this task has hit its previous limit of 10 and failed
the job. In local tests where it was also seen a manual run of the
query shows that the result is '1'. Hence, something is causing this
to take longer, possibly related to patroni but unsure at this time.