Enable proxy routing only on groups where raft is enabled
Per BDR-3039, we should ensure that enable_raft is set for a (sub)group
before we try to set enable_proxy_routing on that group.

Also, we should not disallow enable_proxy_routing on the top-level group
(which was done earlier only in analogy to enable_raft), this turns out
to be a valid thing to do.