TPA-323 Allow for src/install role to more easily be reused
Adds a variable `source_name` to the role, decoupling it from the parent
calling role's use of the `ifs` object. This allows it to be called
from other roles and should enable its reuse, thus cutting down on
code duplication (DRY).

Reduce calls to src/req by moving it up to sys role, so packages are
installed early and attempted only once.