use Cloudsmith repositories for pgd5 installs
Add a new variable edb_repositories, populated at configure-time and
interpreted at deploy-time to add CS repositories.

If we are installing a cluster with bdr_version == 5, we set
edb_repositories to a suitable combination of repositories, and omit the
equivalent repositories from tpa_2q_repositories.

When deploying, we add the listed repositories in the appropriate way
for our OS family, using a token found in edb_repos_token.