TPA-303: Add products/default/release in configure
change the way we set products/default/release for epas.
we previously set it a deploy time. this solution don't allow user to
disable the repo if needed to use private repo which will fail the
deployment.

solution is to add this repo if needed at configure time.
user can then specify an empty tpa_2q_repositories which will be
honored during deploy.

we need to keep the deploy time check to allow backward compatibility.
which explains the slightly more complex check to differentiate empty
and undefined tpa_2q_repositories var.