Exclude _readonly_bdr_roles from CAMO partner eligibility
Commit 62b6967 introduced a subtle regression in the guise of a code
smell cleanup: in creating a _candidate_roles property and "DRY"ing the
PEM/CAMO eligibility tests, it converted both "x and not y and not z and
not …" and "x or y or z …" tests into the same kind of set-intersection
tests against {x,y,z,…}, which inverted the sense of the former test.

The result was that we could end up designating non-primary instances as
CAMO partners, which is obviously incorrect. The problem was not evident
with BDR-Always-ON because its witness nodes (in those layouts which had
any) were (a) odd-numbered, and (b) the last BDR instance, so the CAMO
pairwise assignment never found a partner for them anyway. This is no
longer the case with PGD-Always-ON, hence the bug report.

Here we eliminate _candidate_roles (which, despite seeming plausible,
does not represent any meaningful property of a BDR cluster) in favour
of a _readonly_bdr_roles property and an _is_bdr_primary() test which
is used to determine CAMO eligibility. Thus the _camo_candidate_roles
property can go away.

Separately, _pemagent_candidate_roles can also go away, because it was
effectively testing for "any bdr instance", which we can write in the
obvious way.

References: TPA-342