Introduce `tpaexec reconfigure` and a new lib/tpa class hierarchy
The new tpa classes are centred around a complete object representation
of a config.yml file, comprising a Cluster class with associated pieces
(e.g., Group, Location, and Instance). We can not only write a Cluster
out to a file, but also create one based on an existing file (which is
something the old code had no way to do).

The other basis of the new `reconfigure` command is the Transmogrifier,
which expresses a specific set of changes to a Cluster in a composable
and maintainable way.

Here, we introduce only those changes needed to upgrade a BDR-Always-ON
cluster to PGD-Always-ON, but in principle, `reconfigure` could be made
to change anything about the configuration of an existing cluster.

This commit doesn't touch lib/tpaexec at all, and only reconfigure uses
the new tpa classes. Eventually configure will also be rewritten to use
the the new infrastructure.

With extensive contributions from Jonathan and Richard. Owes its current
form to extensive discussions with Haroon, Giulio, Israel, and Mike.

References: TPA-422