Adapt to BDRv5 CAMO configuration based on commit scopes
In BDRv5, there is no longer any GUC-based configuration for CAMO pairs.
Instead, we must create a CAMO-enforcing commit scope for a subgroup and
set it as the default_commit_scope for the group.

For now, the commit scope is created with the following rule (and you
must edit bdr_commit_scopes in config.yml to set anything different):

    ALL (groupname) ON durable CAMO DEGRADE ON (timeout = 3600s) TO ASYNC

(The timeout in the (optional) DEGRADE ON clause replaces the earlier
bdr.global_commit_timeout and synchronous_replication_availability =
'await' settings.)

Note: BDR was fixed alongside this work to allow setting the default
commit scope for a subgroup. Previously it allowed the default commit
scope to be set only for the topmost group, which is not suitable for
the sort of topology we are working towards here.

References: TPA-342