Handle setting a group's default_commit_scope for BDRv5
If bdr_node_groups[*].options sets `default_commit_scope`, we will now
set it for the relevant group(s) after setting up commit scopes.

Speaking of which, we have to create commit scopes after the group join
stage, because CAMO-enforcing commit scopes want to count the number of
data nodes in the group, so we have to ensure that there ARE data nodes
in the group when we try to create the scope.

Also fix a couple of small problems in 9275b9a, which originally added
commit scope support:

* Fix typo: create_commit_scope → add_commit_scope
* Translate group id from bdr.commit_scopes to a node_group_name via a
  join against bdr.node_group (we need the group name to identify the
  tuple and/or remove the commit scope)

Note: BDRv4 doesn't have bdr.alter_node_group_option(), so we don't
handle it in this commit. We could write equivalent code using the
bdr.alter_node_group_config() function.

Note: bdr.node_group_summary currently contains a "commit_scope" column,
which will be renamed to "default_commit_scope" soon, which means we'll
need to change "catalog_key" in this task.