Adjust bdr extension cleanup condition to match comment/intention
In 8852291, we added a task to remove the BDR extension in databases
where no BDR nodes had been created. This was meant as a cleanup to our
earlier behaviour of creating the bdr extension in template1 and thence
inheriting it into every other database.

Here we tighten up the conditions under which this cleanup is executed:

1. We won't do anything if the bdr extension is not in template1, and
2. We won't do anything for a particular database if bdr is mentioned
   explicitly in postgres_databases[dbname].extensions.

This way, if you create the extension either by hand or by adding bdr to
the extensions list under postgres_databases (for a database other than
bdr_database), it will not be removed even though it is unused at
cleanup time.

This conservative behaviour would not have been unreasonable even in the
initial commit, but now it also removes an additional barrier for custom
hooks that want to configure bdr in databases other than bdr_database.