Remove pgd4 pkg before installing pgd5
As per Florin
on Debian we can use Replace and Breaks on RPM we don't have anything similar.
Obsoletes  goes directly to the newer package(BDR5) even if we want to install bdr4.
So the only way for now is to make them Conflict this way the user needs to
remove bdr4 before installing bdr5