Fail reliably with a useful error if Postgres doesn't start
Investigating a number of recent reports of a failure to look up
cluster_facts.databases for certain hosts in postgres/bdr led us to
realise that check-postgres-available.yml was not failing when Postgres
did not start on some but not all hosts.

(The proximate cause was that we were mistakenly setting bdr.enable_camo
on CAMO-enabled PGD5 clusters, already fixed.)

We used to run postgres-monitor inside a block and use a rescue clause
to display an excerpt from the Postgres logfile. However, although this
is expected to work with `any_errors_fatal`, it actually didn't, and we
couldn't even see the rescue block being executed for failed hosts. One
workaround was to fail the block if any _other_ hosts failed the task,
but that caused all hosts to display a log excerpt (whether they had
failed or not).

Instead of trying to (ahem) rescue this mechanism, it's straightforward
to set ignore_errors on the postgres-monitor task and display the log
excerpt if it failed. Fixes the problem, no known downside.

References: TPA-354