Ensure that nodes subscribe to bdr_child_group, if available
We should include both bdr_node_group and bdr_child_group (if defined)
in the replication sets that a node is expected to be subscribed to by
default (so that we don't go and use alter_node_replication_sets to
unsubscribe to the latter).

Note: this also happens to fix the problem, noticed recently, of
first_bdr_primary having node_join_finished='f' set in bdr.node.

References: TPA-359