Support --active-locations for PGD-Always-ON
By default, we set enable_proxy_routing on the top group. If any active
locations are specified, we set enable_raft+enable_proxy_routing on them
(and no others, not even the top group).

We also remove support for --enable-subgroup-raft, and the mechanism to
set enable_raft on all groups based on the enable_subgroup_raft setting
under cluster_vars (but not the underlying mechanism to set enable_raft
on individual groups). The decision to enable subgroup RAFT should be
based on the features it enables (viz proxy routing to individual
subgroups), and not in the abstract.