Explicitly set platform for docker image in Dockerfile
`FROM` command by default will pull image matching the achitecture of
the machine. If the machine is running using Apple Silicon, the
architecture type will be aarch64 or arm64 instead of amd64, causing
failures in running docker build.

This commit will ensure that we set the platform explicitly to avoid
such failures, or any warnings, such as during `tpaexec info`.