Change HARP Proxy config.yml BDR DCS configuration from single-line to multi-line
This commit basically reverts the changes that were introduced
through `d71fd549da36fbb97f03294bd3cccbd663b0f752`.

That commit was introduced in TPAexec as a workaround for a
HARP issue that was causing connections to HARP proxy to hang.

However, that workaround was causing HARP failovers to take longer
to complete as they would rely on client-side failover checks through
multi-host connection string.

Now that the issue in HARP is addressed we are reverting the changes
in TPAexec so it again deploys HARP proxy with multiple lines for
BDR DCS.

References: TPA-360
References: RT90034

Signed-off-by: Israel Barth <israel.barth@enterprisedb.com>