Explicitly reset {apt,yum}_repository_list for PGD-Always-ON
Even if we want to install community postgresql, the packages should
come from the Cloudsmith repositories (which mirror packages from PGDG).

However, we can only get rid of PGDG, not EPEL, because we still depend
on python3-gunicorn for pg-backup-api. If you don't plan to enable this,
you can get rid of EPEL too.

References: TPA-374