Add a witness node automatically if data_nodes_per_location is even
Earlier, we used to do this only if there were exactly two data nodes.

The --add-witness-node-per-location option is now redundant, since there
is never a reason to specify it, because if we don't do it automatically
anyway, it's disallowed. However, because of the way we count instances
right now, it's easier to not remove it.

References: TPA-371