Translate --edb-postgres-extended → pgextended/edbpge based on architecture
Although we must retain the different postgres_flavour values in
config.yml in order to correctly deploy the two extant variants of
"Postgres Extended", they are distinct in where they can be used, so we
can accept a single configure option (with aliases) and translate that
into a postgres_flavour based on the selected --architecture.

Also add --edb-postgres-advanced as the preferred form of --epas. Here
there are no complications, it's just an alias.