Use ::group:: wrappers to reduce default output
We put all teardown tasks (archive-logs, tar, deprovision) into a single
group, so that the output is de-expanded by default. (Note: it doesn't
seem to work to echo a ::group:: marker in one step and the ::endgroup::
later, so everything has to be in the same step if you want to group it
together.)

Likewise for configure, download-packages, etc.