Fix: Revert forcing docker image platform to amd64
The original commit to set `platform=linux/amd64` was so that it could
run on Apple Silicon Macbooks. However, it is proven to be unnecessary
and build runs take way longer than default arch. This commit will
revert the change and use dynamic architecture for the tpaexec
container.

Debian image will now be using host's architecture and the repository
will be adjusted dynamically with `dpkg` command.

Tested on M1 Max.