Replace --active-locations option with --pgd-proxy-routing
For PGD-Always-ON clusters, proxy routing must be specified.

If --pgd-proxy-routing is set to global, we enable proxy routing
on the top-level group and all proxies are part of the top-level group.
If routing is set to local, we enable proxy routing (and enable_raft)
on the subgroups we create for each location, and proxies are
part of the local subgroup.

This also fixes the bug where pgd-proxies were not added to the
top-level group as expected.

References: TPA-383