Checkpoint to generate WAL for first barman backup
When taking the first barman backup, barman can time out while waiting
for the most recent WAL to be archived. To work around this, we can
generate more WAL so that the WAL we need archived is no longer the most
recent. We connect to the postgres server and call CHECKPOINT to do this.

References: TPA-381