Store the ansible-vault password in the system keyring by default
We used to write the vault password to vault/vault_pass.txt, and we'll
continue to work with existing clusters where that is the case. For new
clusters, we'll try to use the keyring module to store the password into
any available system keyring (e.g., gnome-keyring).

Use `tpaexec configure … --keyring-backend <name>` to select a backend,
but the only available choices are "system" (the default) and "legacy"
(which is to use vault/vault_pass.txt).

Modify use-vault to avoid having to copy it over in cluster_dirs
leverage the fact that tpaexec script `cd` to cluster_dir
before running ansible commands.

References: TPA-85