Refactor CONFIGURE_OPTS to avoid joins and splits
Additionally, add CONFIGURE_EXTRA_ARGS back as it's used by some calling
 workflows.

If the initial CONFIGURE_OPTS declaration has no leading space when a
calling workflow, that supplied an existing arg list without a trailing
space. the two args would be joined.

    CONFIGURE_OPTS="--first-arg"
    CONFIGURE_OPTS+="--architecture foo"

    $CONFIGURE_OPTS == "--first-arg--architecture foo"

Therefore, to avoid word joining like this and to avoid word splitting
(e.g. should an input ever have a space in it), I've converted the
argument construction to use an array.

I've also moved the echo of the final set of args from the end
to before the configure command is run, so that if it fails you get to
see what the string was. As this would put it inside a group which
would hide it from visibility, I've added a new group that splits the
setup of args from the execution of configure.