Fix proxy getting attached to the top-level group
If --pgd-proxy-routing is set to global, we enable proxy routing
on the top-level group and all proxies are part of the top-level group.

However, when proxy routing is enabled on the subgroups we do not
want the proxies to get attached to the top-level group but to their
own local subgroups.

This commit addresses the incorrect behavior where proxies were still
being attached to the top-level group when enable_proxy_routing was
set to true for subgroups.

Reported by Bharat Narayan Telange via TPA-411.