Support provisioning of and deployment to RH9 hosts.
Add RH9 to be accepted by "tpaexec configure". For AWS, add an AMI for
it and make the firstboot script treat it correctly; for docker, add an
image name and enable building of a systemd-enabled image. For all
platforms, make minimal_setup aware of it and fix one package naming
issue.