Avoid using BDR as a product name in the docs
BDR has been rebranded to EDB Postgres Distributed (PGD)
and therefore when we refer to this product or technology
we should use the new branding. PGD itself still uses the
name 'bdr' extensively, for example for the schema.
By this logic it is acceptable that TPA continues to use
'bdr' in variable names such as `bdr_version` but the
documentation should use the correct branding where
referring to PGD in general.