Create a role used to combine object based facts
This is similar in design to the pkg/add_to_list role, except the
result is a deep merged dict-like object. It can be used to construct
object facts in a number of tasks for later use. This creates an
interface which is more readable and easier to debug when it fails.

Previous method would be the use of a combine filter on an object:

    "{{
      object
      |combine(
        (condition)
        |ternary(
          {'true_key': 'value},
          {'false_key': 'value'}
        )
      )
    }}"

With the new role this is simplified to:

  - include_role:
      name: config/add_to_obj
    vars:
      object_varname: object
      when: condition
      object_contents:
        true_key: value
      object_alt_contents:
        false_key: value

The vars `when`, and `object_alt_contents` are optional. If omitted then
the value when the condition is not true is `{}`. The default for `when`
is `true`.

This can also be combined with a `when` guard on the task to prevent
the role include if the condition is false and no alt value is required.
This might save on performance slightly.

  - include_role:
      name: config/add_to_obj
    vars:
      object_varname: object
      object_contents:
        true_key: value
    when: condition