Fix mounted source directories outside of postgres source
Some mounted source directories, e.g. for patroni, are not included
in the mounted_sources list, so the "install from source" method
that uses this attempts to use git clone instead.

Saving a list of all mounted directories is simpler than attempting
to filter and store only those which we're interested in.