Add support for TPA source validation
Since TPA install includes source for playbooks, python and other
scripts, we have seen that an intentional or unintentional attempt to
change the source of the playbooks, python or included scripts can cause
TPA to fail in ways that are extremely difficult to debug.

This change includes a simple verification mechanism as part of `tpaexec
info` command that performs a validation check on TPA installation and
reports if any files fail to match the checksum distributed as part of
TPA package[1]. For source repository clones (strictly allowed for
internal use only), it relies on `git` to validate the source.

This is a weak defence designed to spot changes made without malice. It
is not intended to protects against malicious alterations to the code.
As it only runs as part of `tpaexec info` command, it does not provide
any runtime protection.

[1] Includes a corresponding change in tpaexec-packaging repository as
part of d927f14dbc

References: TPA-322