Use EDB v2 repositories
We need to support `edbpge` which only available in EDB v2 repos.
We would also like to provide Patroni via packages that we build
via foundation.
We do not need to worry about supporting `pgextended` as it is
only available for PDG subscribers and with it only being in 2q
repos it makes the above 2 points difficult to resolve.

Currently, TPA does not automatically add v2 repos due to the
requirement of repmgr in cluster setup and as a default
failover manager. This is likely to change now that replica
creation with `pg_basebackup` is possible, however, to allow for
v2 repos the selection is constrained to `update_cluster_vars`
when patroni as a failover_manager is selected.

Note: raising an error when incompatible combination in
configure is attempted.