Create upstream replica slots before primary slots
We can't allow the upstream replica(s) to start streaming from the
primary in case the downstream replica slot creation gets delayed (for
any reason, not that it would happen in the course of a normal deploy).
So we create the replica slots first, and allow downstream replicas to
start streaming even before there's anything to stream. Only then do we
create slots on the primary.

So, regardless of any concurrent activity on the primary, by the time it
starts sending WAL to its immediate downstream replicas, the cascaded
replicas are already protected against WAL removal on their upstreams.

References: TPA-316