Add support for deploying to SLES 15
New OS family "SUSE" and new OS "SLES" within it.

Add the SUSE family and SLES OS with correct package names and so on to
all the places that are needed for an M1 or PGD-Always-ON cluster
configured with default values.

To support package operations with zypper, we currently shell out to the
command-line client for some operations rather than using ansible's
zypper_repository module, which doesn't currently support a .repo file
containing multiple repositories, such as the PGDG repos provide.

To ensure we don't install the SLES-provided postgres, set PGDG
repositories which we add to have priority over SLES repositories.

Add a docker image for SLES 15.4 and modify the docker support scripts
to add systemd and awk to it.

Add an AWS AMI for SLES 15 and add various SLES-specific actions to the
firstboot script.