Add TPA configure support for Patroni as a failover_manager
Adds the argument `--enable-patroni` and `--patroni-dcs`.

`--patroni-dcs` is a placeholder for future support of alternative
DCS backends, currently only etcd is supported.

HAproxy packages available for RedHat/CentOS 7 are too old to be
supported by TPA. This is mainly due to the age of the openssl library
that comes with the OS.
In order to work around this, TPA with `--enable-patroni` does not
enable HAproxy by default. To use it a new argument `--enable-haproxy`
is needed. Some checks are in place to ensure that the unsupported
configuration is caught both at `configure` and `deploy` time.

References: TPA-392, TPA-396