Add patroni support as failover_manager
Postgres is allowed to complete its standard setup, configuration and
initdb process, then patroni will start and take over management of
the primary node. When more nodes are added they will assume the
replica role and Patroni will perform replica creation.

The M1 cluster layout selected during `configure` consists of 3 etcd
nodes and 2 optional haproxy nodes in addition to the standard M1 nodes.
Source design documentation now publicly accessible on the EDB website,
recommends etcd and haproxy nodes to not run on the same instances as
postgres.

To satisfy the complex requirements for the patroni configuration YAML
file, this has been done
  1) allow for configuration features to be enabled in blocks,
  2) to provide the user with the ability to override any configuration
     parameter or add entire sections that are not present in the
     default block by using `patroni_conf_settings` in config.yml,
  3) and, to render the entire file as YAML therefore getting a free
     validation that it is syntactically valid.

Regarding updates to patroni DCS after cluster init. The configuration
for patroni held in the `bootstrap.dcs` section of the YAML config is
only used once, during the cluster bootstrap/init process. Any changes
needed to the DCS settings need to be submitted to the DCS via either
`patronictl edit-config` or via the /config API. Configuration rendered
in the `roles/patroni/config/tasks/config.yml` file for DCS settings are
compared with the existing DCS settings.

Note, the final settings are compared entirely, so if a user has updated
the DCS outside of TPA the update will always occur. If a change is
required, TPA will issue a `restart` with the `--pending-only` option to
make sure that if patroni detected a change requiring a restart it can
be actioned.

The DCS content is constructed in the following order:

1) Postgres role parameter defaults and user supplied settings
2) Patroni role DCS defaults
3) DCS user settings
4) Patroni feature flags.
5) User settings

Postgres' configuration and its systemd service is managed by default in
TPA. Once patroni is initialised TPA will hand over to Patroni. This
in effect prevents modification of files, such as `pg_hba.conf`,
`pg_ident.conf` and files under the `conf.d` in the data directory. Some
validation steps will raise errors if any incompatible TPA config
options are enabled, including setting a user template for
`postgresql.conf`. This should give the user an opportunity to review
configuration before it is altered by Patroni, if they are migrating an
existing cluster.

The source install method for patroni provides these config options:
    patroni_git_url
    patroni_git_ref
    patroni_src_dir
    patroni_build_dir
    patroni_build_commands

Source install creates a virtual env as the installation target, which
is selected via a prefix var. Existing system python libraries are
incompatible, so this ensures the ecosystem is entirely complete
with the state of the dev requirements.txt.

The source installation is modelled as a development tool as so the
selection of install requirements match that approach.
Patroni requires python3, psycopg and dcs python packages which can be
installed from a number of places, but we opt to get everything and
install the psycopg2 binary as referred in the requirements.dev.txt.
If the developer needs this to be different the patroni_build_commands
can be changed.

When `patroni_ssl_enabled` and SSL certificate is self-signed we need
to include the SSL cert in the requests to the API so that the SSL
client can validate the certificate chain. As ansible does not have
the option to pass this as an argument until `ansible.core` 2.11 we
need to disable validation.

We could also fix this by making sure self-signed certs are included
in the system CA certificates store, so it's added by default. That can
be done in a later commit.

Noting that this is a less than ideal configuration from a security
standpoint and should be fixed before this role can be considered
stable.

References: TPA-392, TPA-398, TPA-399, TPA-400, TPA-397,
            TPA-431, TPA-155, TPA-417

Co-authored-by: Israel <israel.barth@enterprisedb.com>